refactor: autopush.settings -> autopush.config
